requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
astroid==2.4.1
async-timeout==3.0.1
attrs==19.3.0
azure-core==1.5.0
azure-eventhub==5.1.0
azure-eventhub-checkpointstoreblob==1.1.0
azure-eventhub-checkpointstoreblob-aio==1.1.0
azure-identity==1.3.1
black==19.10b0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
idna==2.9
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msal==1.2.0
msal-extensions==0.1.3
msrest==0.6.13
multidict==4.7.5
oauthlib==3.1.0
pathspec==0.8.0
portalocker==1.7.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pywin32==227
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
uamqp==1.2.7
urllib3==1.25.9
wrapt==1.12.1
yarl==1.4.2
